Continue optimizing country scripts

+++ catscripts/Government/Countries/Hungary/Hungary.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-egrep -i 'Hungary' newpages.txt | egrep -iv 'Austria( |-)Hungary' >> Hungary.txt
+KEYWORDS_HUNGARY="Hungary"
+KEYWORDS_HUNGARY_EXCLUDE="Austria( |-)Hungary"
+KEYWORDS_HUNGARY_ALL="$KEYWORDS_HUNGARY"
 
-HUNGARY=$(stat --print=%s Hungary.txt)
-
-if [ $HUNGARY -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Hungary.txt"
-  export CATNAME="Hungary"
-  $CATEGORIZE
-fi
+  
+  debug_start "Hungary"
 
-rm Hungary.txt
+  HUNGARY=$(egrep -i "$KEYWORDS_HUNGARY" newpages.txt | egrep -iv "$KEYWORDS_HUNGARY_EXCLUDE")
+
+  if [ "$HUNGARY" != "" ];
+  then
+    printf "%s" "$HUNGARY" > Hungary.txt
+    export CATFILE="Hungary.txt"
+    export CATNAME="Hungary"
+    $CATEGORIZE
+    rm Hungary.txt
+    unset HUNGARY
+  fi
+
+  debug_end "Hungary"
+
+fi
