+++ catscripts/Government/Countries/Ireland/Ireland.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-egrep -i 'Ireland' newpages.txt | egrep -iv 'Northern Ireland' >> Ireland.txt
+KEYWORDS_IRELAND="Ireland"
+KEYWORDS_IRELAND_EXCLUDE="Northern(| )Ireland"
+KEYWORDS_IRELAND_ALL="$KEYWORDS_IRELAND"
 
-IRELAND=$(stat --print=%s Ireland.txt)
-
-if [ $IRELAND -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Ireland.txt"
-  export CATNAME="Republic of Ireland"
-  $CATEGORIZE
-fi
 
-rm Ireland.txt
+  debug_start "Ireland"
+
+  IRELAND=$(egrep -i "$KEYWORDS_IRELAND" newpages.txt | egrep -iv "$KEYWORDS_IRELAND_EXCLUDE")
+
+  if [ "$IRELAND" != "" ];
+  then
+    printf "%s" "$IRELAND" > Ireland.txt
+    export CATFILE="Ireland.txt"
+    export CATNAME="Republic of Ireland"
+    $CATEGORIZE
+    rm Ireland.txt
+    unset IRELAND
+  fi
+
+  debug_end "Ireland"
+
+fi
