+++ catscripts/Government/Countries/Norway/Norway.sh
@@ -1,19 +1,27 @@
 #!/bin/bash
 
-. ./catscripts/Government/Countries/Denmark/Vikings/Vikings.sh norun
+. ./catscripts/Government/Countries/Denmark/Vikings/Vikings.sh norun #KEYWORDS_VIKINGS
 
 KEYWORDS_NORWAY="Norway"
 KEYWORDS_NORWAY_EXCLUDE="$KEYWORDS_VIKINGS"
+KEYWORDS_NORWAY_ALL="$KEYWORDS_NORWAY"
 
-egrep -i "$KEYWORDS_NORWAY" newpages.txt | egrep -iv "$KEYWORDS_NORWAY_EXCLUDE" > Norway.txt
-
-NORWAY=$(stat --print=%s Norway.txt)
-
-if [ $NORWAY -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Norway.txt"
-  export CATNAME="Norway"
-  $CATEGORIZE
-fi
 
-rm Norway.txt
+  debug_start "Norway"
+
+  NORWAY=$(egrep -i "$KEYWORDS_NORWAY" newpages.txt | egrep -iv "$KEYWORDS_NORWAY_EXCLUDE")
+
+  if [ "$NORWAY" != "" ];
+  then
+    export CATFILE="Norway.txt"
+    export CATNAME="Norway"
+    $CATEGORIZE
+    rm Norway.txt
+    unset NORWAY
+  fi
+
+  debug_end "Norway"
+
+fi
