+++ catscripts/Government/Countries/Switzerland/Switzerland.sh
@@ -5,35 +5,45 @@ KEYWORDS_GENEVA="Geneva"
 KEYWORDS_ZURICH="Zurich"
 KEYWORDS_SWITZERLAND_EXCLUDE="$KEYWORDS_GENEVA|$KEYWORDS_ZURICH"
 
-egrep -i "$KEYWORDS_SWITZERLAND" newpages.txt | egrep -iv "$KEYWORDS_SWITZERLAND_EXCLUDE" >> Switzerland.txt
-egrep -i "$KEYWORDS_GENEVA" newpages.txt >> Geneva.txt
-egrep -i "$KEYWORDS_ZURICH" newpages.txt >> Zurich.txt
-
-SWITZERLAND=$(stat --print=%s Switzerland.txt)
-GENEVA=$(stat --print=%s Geneva.txt)
-ZURICH=$(stat --print=%s Zurich.txt)
-
-if [ $SWITZERLAND -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Switzerland.txt"
-  export CATNAME="Switzerland"
-  $CATEGORIZE
-fi
+  
+  debug_start "Switzerland"
 
-if [ $GENEVA -ne 0 ];
-then
-  export CATFILE="Geneva.txt"
-  export CATNAME="Geneva"
-  $CATEGORIZE
-fi
+  SWITZERLAND=$(egrep -i "$KEYWORDS_SWITZERLAND" newpages.txt | egrep -iv "$KEYWORDS_SWITZERLAND_EXCLUDE")
+  GENEVA=$(egrep -i "$KEYWORDS_GENEVA" newpages.txt)
+  ZURICH=$(egrep -i "$KEYWORDS_ZURICH" newpages.txt)
 
-if [ $ZURICH -ne 0 ];
-then
-  export CATFILE="Zurich.txt"
-  export CATNAME="Zurich"
-  $CATEGORIZE
-fi
+  if [ "$SWITZERLAND" != "" ];
+  then
+    printf "%s" "$SWITZERLAND" > Switzerland.txt
+    export CATFILE="Switzerland.txt"
+    export CATNAME="Switzerland"
+    $CATEGORIZE
+    rm Switzerland.txt
+    unset SWITZERLAND
+  fi
 
-rm Switzerland.txt
-rm Geneva.txt
-rm Zurich.txt
+  if [ "$GENEVA" != "" ];
+  then
+    printf "%s" "$GENEVA" > Geneva.txt
+    export CATFILE="Geneva.txt"
+    export CATNAME="Geneva"
+    $CATEGORIZE
+    rm Geneva.txt
+    unset GENEVA
+  fi
+
+  if [ "$ZURICH" != "" ];
+  then
+    printf "%s" "$ZURICH" > Zurich.txt
+    export CATFILE="Zurich.txt"
+    export CATNAME="Zurich"
+    $CATEGORIZE
+    rm Zurich.txt
+    unset ZURICH
+  fi
+
+  debug_end "Switzerland"
+
+fi
